pyproject.toml
[project]
name = "steampulse-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bcrypt>=5.0.0",
    "fastapi>=0.117.1",
    "jwt>=1.4.0",
    "mercadopago>=2.3.0",
    "motor>=3.7.1",
    "pydantic[email]>=2.11.9",
    "python-dotenv>=1.1.1",
    "python-jose[cryptography]>=3.5.0",
    "uvicorn[standard]>=0.37.0",
]
